Homework 1 Screenshot
I can't use Virtual Box as it is not supported with Apples silicone processor. I am instead using a paid for service that I already had called Parallels which does the same thing and more. I show that I have ubuntu desktop and a terminal to work out of. I also show my Mac terminal as I plan to use that as well.